Refactor temple to use a "forge" for ls and other remote commands
Temple previously assumed that only Github was used as a git forge. This
refactors the code to instantiate a git forge based on the path.

For example, users now type ``temple ls github.com/User`` to list
templates from a Github forge. Similarly, a forge is used to obtain
the latest template version.

This refactor allows for a clean integration with Gitlab.
