--- docs/future_work.rst
@@ -1,4 +1,0 @@
-Future Work
-===========
-
-1. Use python wrappers for git and Github access (`GitPython <http://gitpython.readthedocs.io/en/stable/>`_, `github3.py <https://github3py.readthedocs.io/en/master/>`_)
